Proposal for refinement of Vincenty's formula
If the distance between the two points is zero, the squared distance is also zero. 
This of course only has any relevance in this single case, so very minor

Disclaimer: all test were done in C#.
I couldn't find any pair of coordinates where `isNan(cos2σM)` wasn't triggered by `cosSqα == 0`.
It that really is the case, at least I find it more readable to check the divisor for 0 before performing the division instead of checking for NaN afterwards.

Thanks for all the hard work, it is very appreciated!