Refactor eq'n (6) to take full advantage of its performant design
From the original paper:

    The compactness of the recommended solutions is due to the use of nested
equations to compute elliptic terms and of only three trigonometric functions:
sine, cosine, and arc tangent.  Nesting reduces the number of operations involving
storage and retrieval of intermediate results (particularly when programming)
in assembly language), reduces the length of the program and the time of execution,
and minimizes the possibility of underflow.

TODO: benchmarking and performance analysis to validate that this change
is appropriately taken advantage of by modern engines such as V8 and
Spidermonkey